refactor: 275 - extract widgets
