+++ lib/pages/cej_information_page.dart
@@ -11,6 +11,7 @@ import 'package:pass_emploi_app/widgets/primary_action_button.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
 
 import '../ui/strings.dart';
+import '../widgets/cej_information_content_card.dart';
 import 'choix_organisme_page.dart';
 
 class CejInformationPage extends StatefulWidget {
@@ -73,9 +74,9 @@ class _CejInformationPageState extends State<CejInformationPage> {
                   child: PageView(
                     controller: _controller,
                     children: [
-                      _firstCard(),
-                      _secondCard(),
-                      _thirdCard(),
+                      const CejInformationFirstContentCard(),
+                      const CejInformationSecondContentCard(),
+                      const CejInformationThirdContentCard()
                     ],
                   ),
                 ),
@@ -130,113 +131,6 @@ class _CejInformationPageState extends State<CejInformationPage> {
       onPressed: () {
         Navigator.of(context).pop();
       },
-    );
-  }
-
-  Widget _contentCard(List<Widget> children) {
-    return Padding(
-      padding: const EdgeInsets.all(Margins.spacing_m),
-      child: Container(
-        decoration: BoxDecoration(color: Colors.white, borderRadius: BorderRadius.circular(8)),
-        child: SingleChildScrollView(
-          child: Padding(
-            padding: const EdgeInsets.all(Margins.spacing_m),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: children,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _firstCard() {
-    return _contentCard(
-      [
-        Text(Strings.whatIsPassEmploi, style: TextStyles.textMBold.copyWith(color: AppColors.primary)),
-        SizedBox(height: Margins.spacing_base),
-        Text(Strings.whatIsPassEmploiDesc, style: TextStyles.textBaseRegular),
-        SizedBox(height: Margins.spacing_m),
-        _cardBulletPoint(Drawables.icDone, Strings.customService),
-        SizedBox(height: Margins.spacing_s),
-        Text(Strings.customServiceDesc, style: TextStyles.textBaseRegular),
-        SizedBox(height: Margins.spacing_m),
-        _cardBulletPoint(Drawables.icOnboardingChat, Strings.favoredContact),
-        SizedBox(height: Margins.spacing_s),
-        Text(Strings.favoredContactDesc, style: TextStyles.textBaseRegular),
-        SizedBox(height: Margins.spacing_m),
-        _cardBulletPoint(Drawables.icSearch, Strings.searchTool),
-        SizedBox(height: Margins.spacing_s),
-        Text(Strings.searchToolDesc, style: TextStyles.textBaseRegular),
-      ],
-    );
-  }
-
-  Widget _secondCard() {
-    return _contentCard(
-      [
-        Text(Strings.whatIsCej, style: TextStyles.textMBold.copyWith(color: AppColors.primary)),
-        SizedBox(height: Margins.spacing_m),
-        _cardBulletPoint(Drawables.icFlash, Strings.customService),
-        SizedBox(height: Margins.spacing_s),
-        Text(Strings.customServiceCejDesc, style: TextStyles.textBaseRegular),
-        SizedBox(height: Margins.spacing_m),
-        _cardBulletPoint(Drawables.icPeople, Strings.uniqueReferent),
-        SizedBox(height: Margins.spacing_s),
-        Text(Strings.uniqueReferentDesc, style: TextStyles.textBaseRegular),
-        SizedBox(height: Margins.spacing_m),
-        _cardBulletPoint(Drawables.icEuro, Strings.financialHelp),
-        SizedBox(height: Margins.spacing_s),
-        Text(Strings.financialHelpDesc, style: TextStyles.textBaseRegular),
-      ],
-    );
-  }
-
-  Widget _thirdCard() {
-    return _contentCard([
-      Text(Strings.whoIsConcerned, style: TextStyles.textMBold.copyWith(color: AppColors.primary)),
-      SizedBox(height: Margins.spacing_m),
-      SvgPicture.asset(Drawables.puzzle),
-      SizedBox(height: Margins.spacing_m),
-      RichText(
-        text: TextSpan(
-          children: [
-            TextSpan(text: Strings.whoIsConcernedFirstRichText[0], style: TextStyles.textBaseRegular),
-            TextSpan(text: Strings.whoIsConcernedFirstRichText[1], style: TextStyles.textBaseBold),
-            TextSpan(text: Strings.whoIsConcernedFirstRichText[2], style: TextStyles.textBaseRegular),
-            TextSpan(text: Strings.whoIsConcernedFirstRichText[3], style: TextStyles.textBaseBold),
-            TextSpan(text: Strings.whoIsConcernedFirstRichText[4], style: TextStyles.textBaseRegular),
-          ],
-        ),
-      ),
-      SizedBox(height: Margins.spacing_m),
-      RichText(
-        text: TextSpan(
-          children: [
-            // → Les personnes qui ne sont pas en formation ni en emploi durable (CDI ou CDD de longue durée)
-            TextSpan(text: Strings.whoIsConcernedSecondRichText[0], style: TextStyles.textBaseRegular),
-            TextSpan(text: Strings.whoIsConcernedSecondRichText[1], style: TextStyles.textBaseBold),
-            TextSpan(text: Strings.whoIsConcernedSecondRichText[2], style: TextStyles.textBaseRegular),
-          ],
-        ),
-      ),
-    ]);
-  }
-
-  Widget _cardBulletPoint(String icon, String text) {
-    return RichText(
-      text: TextSpan(
-        children: [
-          WidgetSpan(
-            child: Padding(
-              padding: const EdgeInsets.only(right: Margins.spacing_s),
-              child: SvgPicture.asset(icon),
-            ),
-          ),
-          TextSpan(text: text, style: TextStyles.textBaseBold),
-        ],
-      ),
     );
   }
 }
